requirements/base.txt
Django==4.1.7
environs[django]==9.5.0
djangorestframework==3.14.0
psycopg2-binary==2.9.5
drf-spectacular==0.25.1
djangorestframework-simplejwt==5.2.2
django-cors-headers==3.13.0
parameterized==0.8.1